requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
emailmessage==0.2.1
et-xmlfile==1.1.0
flask==3.0.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.2
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.4
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
